# docs/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/comments": {
            "post": {
                "operationId": "CommentController_createComment",
                "summary": "发表评论",
                "description": "发表评论,需要注册用户权限",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCommentBodyDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "发表成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateCommentResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    }
                },
                "tags": [
                    "评论"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/comments/by-area/{id}": {
            "get": {
                "operationId": "CommentController_getCommentsByAreaID",
                "summary": "通过评论区ID获取评论列表",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "评论区ID",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "rootLimit",
                        "required": true,
                        "in": "query",
                        "description": "分页参数,一级评论限制数量",
                        "schema": {
                            "minimum": 1,
                            "maximum": 30,
                            "type": "number"
                        }
                    },
                    {
                        "name": "childrenLimit",
                        "required": true,
                        "in": "query",
                        "description": "分页参数,二级评论限制数量",
                        "schema": {
                            "minimum": 1,
                            "maximum": 30,
                            "type": "number"
                        }
                    },
                    {
                        "name": "rootStart",
                        "required": true,
                        "in": "query",
                        "description": "分页参数,一级评论开始位置",
                        "schema": {
                            "minimum": 0,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetCommentsByAreaIDResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    }
                },
                "tags": [
                    "评论"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/comments/by-parent/{id}": {
            "get": {
                "operationId": "CommentController_getCommentsByParentID",
                "summary": "通过父评论ID获取评论列表",
                "description": "通过一级评论ID获取对应的子评论列表",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "父评论ID",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "required": true,
                        "in": "query",
                        "description": "分页参数,限制数量",
                        "schema": {
                            "minimum": 1,
                            "maximum": 30,
                            "type": "number"
                        }
                    },
                    {
                        "name": "start",
                        "required": true,
                        "in": "query",
                        "description": "分页参数,开始位置",
                        "schema": {
                            "minimum": 0,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GetCommentsByParentIDResponseDto"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    }
                },
                "tags": [
                    "评论"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/comments/by-conversation/{id}": {
            "get": {
                "operationId": "CommentController_getCommentsByConversationID",
                "summary": "通过讨论ID获取评论列表",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "讨论ID",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "required": true,
                        "in": "query",
                        "description": "分页参数,限制数量",
                        "schema": {
                            "minimum": 1,
                            "maximum": 30,
                            "type": "number"
                        }
                    },
                    {
                        "name": "start",
                        "required": true,
                        "in": "query",
                        "description": "分页参数,开始位置",
                        "schema": {
                            "minimum": 0,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GetCommentsByConversationIDResponseDto"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    }
                },
                "tags": [
                    "评论"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/comment/{id}": {
            "delete": {
                "operationId": "CommentController_deleteComment",
                "summary": "删除评论",
                "description": "删除评论,需要注册用户权限",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "评论ID",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteCommentResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    }
                },
                "tags": [
                    "评论"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/manage/cat/{id}": {
            "put": {
                "operationId": "CatController_updateCat",
                "summary": "修改猫咪信息",
                "description": "修改猫咪信息,需要管理员权限",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "猫咪ID",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCatBodyDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateCatResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    }
                },
                "tags": [
                    "管理"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "operationId": "CommentController_deleteComment",
                "summary": "删除评论",
                "description": "删除评论,需要管理员权限",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "评论ID",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteCommentResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    }
                },
                "tags": [
                    "管理"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/manage/comments": {
            "get": {
                "operationId": "CommentController_getPhotos",
                "summary": "获取所有评论",
                "description": "获取所有评论,需要管理员权限",
                "parameters": [
                    {
                        "name": "limit",
                        "required": true,
                        "in": "query",
                        "description": "分页参数,限制数量",
                        "schema": {
                            "minimum": 1,
                            "maximum": 30,
                            "type": "number"
                        }
                    },
                    {
                        "name": "start",
                        "required": true,
                        "in": "query",
                        "description": "分页参数,开始位置",
                        "schema": {
                            "minimum": 0,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GetCommentsResponseDto"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    }
                },
                "tags": [
                    "管理"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "Kitty后端APi服务",
        "description": {
            "$ref": "./index.md"
        },
        "version": "1.15",
        "contact": {},
        "x-logo": {
            "url": "./logo.png",
            "backgroundColor": "#FFFFFF"
        }
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "name": "统一认证",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateCommentBodyDto": {
                "type": "object",
                "properties": {
                    "areaID": {
                        "type": "number",
                        "description": "评论区ID",
                        "minimum": 1
                    },
                    "parentID": {
                        "type": "number",
                        "description": "父评论ID",
                        "minimum": 1
                    },
                    "conversationID": {
                        "type": "number",
                        "description": "讨论ID",
                        "minimum": 1
                    },
                    "content": {
                        "type": "string",
                        "description": "评论内容"
                    }
                },
                "required": [
                    "areaID",
                    "content"
                ]
            },
            "CreateCommentResponseDto": {
                "type": "object",
                "properties": {}
            },
            "GetCommentsByAreaIDResponseDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "description": "评论ID",
                        "minimum": 1
                    },
                    "conversationID": {
                        "type": "number",
                        "description": "对话ID",
                        "minimum": 1
                    },
                    "content": {
                        "type": "string",
                        "description": "评论内容"
                    },
                    "createdDate": {
                        "format": "date-time",
                        "type": "string",
                        "description": "发表时间"
                    },
                    "childrenTotal": {
                        "type": "number",
                        "description": "子评论总数",
                        "minimum": 0
                    },
                    "childrenComments": {
                        "description": "子评论",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "id",
                    "conversationID",
                    "content",
                    "createdDate",
                    "childrenTotal",
                    "childrenComments"
                ]
            },
            "GetCommentsByParentIDResponseDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "description": "评论ID",
                        "minimum": 1
                    },
                    "conversationID": {
                        "type": "number",
                        "description": "对话ID",
                        "minimum": 1
                    },
                    "content": {
                        "type": "string",
                        "description": "评论内容"
                    },
                    "createdDate": {
                        "format": "date-time",
                        "type": "string",
                        "description": "发表时间"
                    }
                },
                "required": [
                    "id",
                    "conversationID",
                    "content",
                    "createdDate"
                ]
            },
            "GetCommentsByConversationIDResponseDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "description": "评论ID",
                        "minimum": 1
                    },
                    "conversationID": {
                        "type": "number",
                        "description": "对话ID",
                        "minimum": 1
                    },
                    "content": {
                        "type": "string",
                        "description": "评论内容"
                    },
                    "createdDate": {
                        "format": "date-time",
                        "type": "string",
                        "description": "发表时间"
                    }
                },
                "required": [
                    "id",
                    "conversationID",
                    "content",
                    "createdDate"
                ]
            },
            "DeleteCommentResponseDto": {
                "type": "object",
                "properties": {}
            },
...
            "UpdateCatBodyDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "猫咪名称"
                    },
                    "species": {
                        "type": "string",
                        "description": "猫咪物种"
                    },
                    "isNeuter": {
                        "type": "boolean",
                        "description": "猫咪是否已经绝育"
                    },
                    "description": {
                        "type": "string",
                        "description": "猫咪描述"
                    },
                    "haunt": {
                        "type": "string",
                        "description": "猫咪出没地点"
                    },
                    "status": {
                        "type": "string",
                        "description": "猫咪状态,0表示失踪,1表示正常,2表示生病,3表示住院,4表示被领养,5表示死亡",
                        "enum": [
                            "0",
                            "1",
                            "2",
                            "3",
                            "4",
                            "5"
                        ]
                    }
                },
                "required": [
                    "name",
                    "species",
                    "isNeuter",
                    "description",
                    "haunt",
                    "status"
                ]
            },
            "UpdateCatResponseDto": {
                "type": "object",
                "properties": {}
            },
...
            "GetCommentsResponseDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "description": "评论ID",
                        "minimum": 1
                    },
                    "conversationID": {
                        "type": "number",
                        "description": "对话ID",
                        "minimum": 1
                    },
                    "content": {
                        "type": "string",
                        "description": "评论内容"
                    },
                    "status": {
                        "type": "string",
                        "description": "评论状态,0表示未审核,1表示正常显示,2表示隐藏"
                    },
                    "userID": {
                        "type": "number",
                        "description": "照片发布者ID"
                    },
                    "parentCommentID": {
                        "type": "number",
                        "description": "父评论ID"
                    },
                    "areaID": {
                        "type": "number",
                        "description": "评论区ID"
                    },
                    "createdDate": {
                        "format": "date-time",
                        "type": "string",
                        "description": "发表时间"
                    }
                },
                "required": [
                    "id",
                    "conversationID",
                    "content",
                    "status",
                    "userID",
                    "parentCommentID",
                    "areaID",
                    "createdDate"
                ]
            },
...
            "HttpExceptionResponseDto": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "number",
                        "description": "HTTP状态码"
                    },
                    "msg": {
                        "type": "object",
                        "description": "错误信息"
                    }
                },
                "required": [
                    "status",
                    "msg"
                ]
            }
        },
        "responses": {
            "400": {
                "description": "参数错误",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/HttpExceptionResponseDto"
                        }
                    }
                }
            },
            "401": {
                "description": "Authorization字段无效",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/HttpExceptionResponseDto"
                        }
                    }
                }
            },
            "403": {
                "description": "无权限",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/HttpExceptionResponseDto"
                        }
                    }
                }
            }
        }
    }
}